revision of posts model and adding pagination

+++ backend/src/main/java/com/seproject/Bookface/post/PostServiceImpl.java
@@ -2,14 +2,17 @@ package com.seproject.Bookface.post;
 
 import com.seproject.Bookface.post.dao.PostEntity;
 import com.seproject.Bookface.post.dto.request.CreatePostRequest;
-import com.seproject.Bookface.post.dto.response.AllPostsFromUserResponse;
+import com.seproject.Bookface.post.dto.response.PostsResponse;
 import com.seproject.Bookface.user.dto.response.MeResponse;
+import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.sql.Timestamp;
 import java.time.LocalDateTime;
-import java.util.Optional;
+import java.util.Objects;
 
 @Service
 public class PostServiceImpl implements PostService {
@@ -21,38 +24,71 @@ public class PostServiceImpl implements PostService {
     }
 
     @Override
-    public void addPost(CreatePostRequest requestBody) {
+    public ResponseEntity<String> addPost(CreatePostRequest requestBody, String userId) {
         MeResponse me = (MeResponse) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        String userId = me.getUserId();
+        String auth = me.getUserId();
 
-        postRepository.save(PostEntity.builder()
-                .userId(userId)
-                .title(requestBody.getTitle())
-                .content(requestBody.getContent())
-                .timestamp(Timestamp.valueOf(LocalDateTime.now()))
-                .build());
+        if (Objects.equals(userId, auth)) {
+            postRepository.save(PostEntity.builder()
+                    .userId(userId)
+                    .title(requestBody.getTitle())
+                    .content(requestBody.getContent())
+                    .timestamp(Timestamp.valueOf(LocalDateTime.now()))
+                    .build());
+            return new ResponseEntity<>("Post successfully added", HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>("Unauthorized access", HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @Override
-    public void removePost(Long postId) {
-        postRepository.deleteById(postId);
+    public ResponseEntity<String> removePost(String postId, String userId) {
+        MeResponse me = (MeResponse) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String auth = me.getUserId();
+
+        if (Objects.equals(userId, auth)) {
+            if (postRepository.existsById(postId)) {
+                postRepository.deleteById(postId);
+                return new ResponseEntity<>("Post successfully deleted", HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>("Post not found", HttpStatus.BAD_REQUEST);
+            }
+        } else {
+            return new ResponseEntity<>("Unauthorized access", HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @Override
-    public Optional<PostEntity> modifyPost(Long postId, String title, String content) {
-        postRepository.setUserInfoById(postId, title, content);
+    public ResponseEntity<String> modifyPost(String postId, String title, String content, String userId) {
+        MeResponse me = (MeResponse) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String auth = me.getUserId();
 
-        return postRepository.findById(postId);
+        if (Objects.equals(userId, auth)) {
+            if (postRepository.existsById(postId)) {
+                postRepository.setUserInfoById(postId, title, content);
+                return new ResponseEntity<>("Post successfully modified", HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>("Post not found", HttpStatus.BAD_REQUEST);
+            }
+        } else {
+            return new ResponseEntity<>("Unauthorized access", HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @Override
-    public Optional<PostEntity> findPost(Long postId) {
-        return postRepository.findById(postId);
+    public ResponseEntity<PostEntity> getPost(String userId, String postId) {
+        if (postRepository.existsById(postId)) {
+            if (Objects.equals(postRepository.getPostEntityByPostId(postId).getUserId(), userId)) {
+                return new ResponseEntity<>(postRepository.getPostEntityByPostId(postId), HttpStatus.OK);
+            }
+        }
+        return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
     }
 
     @Override
-    public AllPostsFromUserResponse findAllPostsFromUser(String userId) {
-        AllPostsFromUserResponse allPosts = new AllPostsFromUserResponse(postRepository.findAllByUserId(userId));
+    public PostsResponse findAllPostsFromUser(String userId, Pageable paging) {
+        PostsResponse allPosts = new PostsResponse(postRepository
+                .findAllByUserIdOrderByTimestampDesc(userId, paging).getContent());
         return allPosts;
     }
 }
